+++ routes/auth.js
@@ -1,0 +1,10 @@
+const express = require("express");
+const router = express.Router();
+const authController = require("../controllers/auth");
+const auth = require("../middlewares/auth");
+
+router.post("/register", authController.register);
+
+router.post("/login", authController.login, [auth.authToken]);
+
+module.exports = router;
