+++ routes/medicalRecord.js
@@ -6,4 +6,6 @@ router.post("/create", medicalRecordController.createMedicalRecord);
 
 router.patch("/edit/:id", medicalRecordController.updateMedicalRecord);
 
+router.get("/:id", medicalRecordController.getMRById);
+
 module.exports = router;
