+++ routes/user.js
@@ -5,11 +5,9 @@ const auth = require("../middlewares/auth");
 
 router.get("/", userController.getUsers);
 
-router.post("/register", userController.register);
-
-router.post("/login", userController.login, [auth.authToken]);
-
 router.patch("/editProfile/:id", userController.editProfile);
+
+router.get("/:id", userController.getUserById);
 
 // router.post("/create-profile", userController.userCreate);
